+++ jopa-integration-tests-jena/src/test/java/cz/cvut/kbss/jopa/test/query/jena/TypedQueryTest.java
@@ -47,6 +47,7 @@ public class TypedQueryTest extends TypedQueryRunner {
         final Map<String, String> properties = new HashMap<>();
         properties.put(JenaOntoDriverProperties.JENA_STORAGE_TYPE, JenaOntoDriverProperties.IN_MEMORY);
         properties.put(JenaOntoDriverProperties.JENA_TREAT_DEFAULT_GRAPH_AS_UNION, Boolean.toString(true));
+        properties.put(JenaOntoDriverProperties.JENA_ISOLATION_STRATEGY, JenaOntoDriverProperties.SNAPSHOT);
         em = persistenceFactory.getEntityManager("SPARQLTypedQueryTests", false, properties);
         QueryTestEnvironment.generateTestData(em);
         em.clear();