+++ modelgen/src/main/java/cz/cvut/kbss/jopa/modelgen/OutputFilesGenerator.java
@@ -27,7 +27,6 @@ import javax.annotation.processing.Messager;
 import javax.tools.Diagnostic;
 import java.io.File;
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.StandardOpenOption;
 import java.util.Collection;
@@ -75,8 +74,7 @@ public class OutputFilesGenerator {
         content.append(generateAttributes(cls));
         content.append(generateClassSuffix());
         try {
-            Files.write(targetFile.toPath(), content.toString()
-                                                    .getBytes(StandardCharsets.UTF_8), StandardOpenOption.TRUNCATE_EXISTING);
+            Files.writeString(targetFile.toPath(), content.toString(), StandardOpenOption.TRUNCATE_EXISTING);
         } catch (IOException e) {
             throw new ModelGenException("Unable to output content to target file '" + targetFile + "'!", e);
         }
@@ -155,75 +153,58 @@ public class OutputFilesGenerator {
         StringBuilder attributes = new StringBuilder();
         for (Field field : cls.getFields()) {
             final String declaringClass = field.getParentName().substring(field.getParentName().lastIndexOf('.') + 1);
-            attributes.append("\t public static volatile ");
+            attributes.append("    public static volatile ");
             //@Id
             if (isAnnotatedWith(field, MappingAnnotations.ID)) {
-                attributes
-                        .append("Identifier<")
-                        .append(declaringClass)
-                        .append(", ")
-                        .append(field.getType().getTypeName()
-                                     .substring(field.getType().getTypeName().lastIndexOf(".") + 1));
+                attributes.append("Identifier<")
+                          .append(declaringClass)
+                          .append(", ")
+                          .append(field.getType().getTypeName()
+                                       .substring(field.getType().getTypeName().lastIndexOf(".") + 1));
                 //@Types
             } else if (isAnnotatedWith(field, MappingAnnotations.TYPES)) {
-                attributes
-                        .append("TypesSpecification<")
-                        .append(declaringClass)
-                        .append(", ");
+                attributes.append("TypesSpecification<")
+                          .append(declaringClass)
+                          .append(", ");
                 if (field.getType().getIsSimple()) {
-                    attributes
-                            .append(field.getType().getTypeName()
-                                         .substring(field.getType().getTypeName().lastIndexOf(".") + 1));
+                    attributes.append(field.getType().getSimpleName());
                 } else {
-                    attributes
-                            .append(field.getType().getTypes().get(0).getTypeName()
-                                         .substring(field.getType().getTypes().get(0).getTypeName()
-                                                         .lastIndexOf(".") + 1));
+                    attributes.append(field.getType().getTypes().get(0).getSimpleName());
                 }
                 //@Properties
             } else if (isAnnotatedWith(field, MappingAnnotations.PROPERTIES)) {
-                attributes
-                        .append("PropertiesSpecification<")
-                        .append(declaringClass)
-                        .append(", ");
+                attributes.append("PropertiesSpecification<")
+                          .append(declaringClass)
+                          .append(", ");
                 Type type = field.getType();
                 if (!Objects.equals(type.getTypeName(), Map.class.getName())) {
                     attributes
                             .append(type.getTypes().get(0).getTypeName()
                                         .substring(type.getTypes().get(0).getTypeName().lastIndexOf(".") + 1));
                 } else {
-                    attributes
-                            .append("Map, ")
-                            .append(type.getTypes().get(0).getTypeName()
-                                        .substring(type.getTypes().get(0).getTypeName().lastIndexOf(".") + 1))
-                            .append(", ")
-                            .append(type.getTypes().get(1).getTypes().get(0).getTypeName()
-                                        .substring(type.getTypes().get(1).getTypes().get(0).getTypeName()
-                                                       .lastIndexOf(".") + 1));
+                    attributes.append("Map, ")
+                              .append(type.getTypes().get(0).getSimpleName())
+                              .append(", ")
+                              .append(type.getTypes().get(1).getTypes().get(0).getSimpleName());
                 }
             } else if (isAnnotatedWith(field, MappingAnnotations.DATA_PROPERTY)
                     || isAnnotatedWith(field, MappingAnnotations.OBJECT_PROPERTY)
                     || isAnnotatedWith(field, MappingAnnotations.ANNOTATION_PROPERTY)) {
                 Type type = field.getType();
                 if (type.getIsSimple()) {
-                    attributes
-                            .append("SingularAttribute<")
-                            .append(declaringClass)
-                            .append(", ")
-                            .append(type.getTypeName().substring(type.getTypeName().lastIndexOf(".") + 1));
+                    attributes.append("SingularAttribute<")
+                              .append(declaringClass)
+                              .append(", ")
+                              .append(type.getSimpleName());
                 } else {
                     if (type.getTypeName().equals(List.class.getName())) {
-                        attributes
-                                .append("ListAttribute<");
+                        attributes.append("ListAttribute<");
                     } else if (type.getTypeName().equals(Set.class.getName())) {
-                        attributes
-                                .append("SetAttribute<");
+                        attributes.append("SetAttribute<");
                     }
-                    attributes
-                            .append(declaringClass)
-                            .append(", ")
-                            .append(type.getTypes().get(0).getTypeName()
-                                        .substring(type.getTypes().get(0).getTypeName().lastIndexOf(".") + 1));
+                    attributes.append(declaringClass)
+                              .append(", ")
+                              .append(type.getTypes().get(0).getSimpleName());
                 }
             }
             attributes